LayoutTests/accessibility/text-marker/media-emits-object-replacement-expected.txt


This tests ensures that if video or audio tags are used, they will emit an object replacement character in a range for string operation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text1.textMarkerRangeLength(textMarkerRange) is 5
Video string for range: a ￼ b
PASS text1.textMarkerRangeLength(textMarkerRange) is 5
Audio string for range: a ￼ b
PASS successfullyParsed is true

TEST COMPLETE
